pipeline/requirements-dev.txt

isort
black
flake8
mypy
coverage
docformatter

pandas-stubs
types-google-cloud-ndb
django-stubs
pytest
pytest-django
pytest-cov